requirements.txt
fastapi
uvicorn
nltk
wikipedia
requests
python-dotenv
jinja2
python-multipart
